chore(build): retire the prose-closeout machinery, keep the recurring check
The 472-product provenance closeout is finished and tagged (baseline-472-2026-08-16).
What it left behind was a migration toolchain sitting in `sources/` and `build/` as
though it were live: 16 decision manifests plus a 472-row ledger under `sources/`,
which reads as a build input and is not, and three appliers whose work is done.

Removed:

  sources/provenance/            16 manifests + LEDGER.yaml
  build/apply_provenance.py      wrote the derive rows
  build/repair_placeholder_shows.py + its digest file
  build/migrate_banded_quantity.py
  tests/test_prose_provenance.py

Kept, as build/product_prose.py: the five-state classifier, the verification-date
extraction sweep_status ages prose on, and METHOD_WORDS. Those are not migration
code — a new or edited product can reintroduce any unresolved state, and the
classifier is what notices.

Dropped from it: packet generation, candidate gathering, the clause-rewrite
machinery, and the clause-boundary heuristic, none of which have a caller now.

The surviving tests moved into tests/test_product_prose.py, where they join the
marketing and thin-description checks on the same corpus. The corpus-wide
invariant is now asserted directly — every product canonical — rather than as a
census the closeout read by hand.

sibling-invariants.md loses the three entries that described the applier itself.
Its rule stands; the modules it caught the defect in are gone, and git has the
detail. check_payload's date entry keeps its half of the shared regression test.

609 tests pass, no skips. validate, check_verification, check_freshness,
check_payload, check_routing and check_retirement unchanged.
