+++ docs/reference/sibling-invariants.md
@@ -12,8 +12,8 @@ it claimed to answer**:
 |---|---|
 | `check_routing.SOURCE_ARTIFACT` named five sources; the table declared seven | "is this one of the five sources I know about?" |
 | `check_routing.artifacts_of` enumerated the same vocabulary again, in the same file | "does the product declare one of seven hardcoded keys?" |
-| `apply_provenance` reimplemented `METHOD_WORDS` more narrowly | "is this document name one of three forbidden phrases?" |
-| `check_payload` and `apply_provenance` each validated a date by shape | "does this string look like a date?" |
+| a second, narrower `METHOD_WORDS` in the prose applier | "is this document name one of three forbidden phrases?" |
+| two modules each validated a date by shape | "does this string look like a date?" |
 
 Three were found by review rather than by any gate, and the fourth was found only because the
 third was. The pattern is not carelessness about a particular constant — it is that a fix gets
@@ -37,7 +37,6 @@ and tested, so it stays a decision instead of becoming drift.
     - build/sweep_status.py
     - build/check_payload.py
     - build/validate.py          # inline, in the schema gate itself
-    - build/repair_placeholder_shows.py  # inline
     - build/render.py            # inline
   risk: >-
     A fourth axis narrows ten denominators at once, and a walk that quietly skips an axis
@@ -80,9 +79,9 @@ and tested, so it stays a decision instead of becoming drift.
   regression_test: test_capability_relation_deltas_match_the_schema_enum
 
 - construct: vocabulary — the method words a provenance line may not name
-  canonical_owner: build/prose_provenance.py METHOD_WORDS
+  canonical_owner: build/product_prose.py METHOD_WORDS
   duplicates:
-    - build/apply_provenance.py   # fixed in #283
+    - the prose applier   # fixed in #283, module retired in the 2026-08 cleanup
   risk: >-
     The sibling was narrower and let `substitute sources` through as a document name — the
     exact phrase behind four hardware records that a rewording pass nearly promoted into
@@ -94,64 +93,18 @@ and tested, so it stays a decision instead of becoming drift.
   canonical_owner: datetime.date.fromisoformat
   duplicates:
     - build/check_payload.py      # fixed
-    - build/apply_provenance.py   # fixed in #283
+    - the prose applier           # fixed in #283, module retired
   risk: >-
     A `\d{4}-\d{2}-\d{2}` regex accepts 2026-99-99 and 2026-02-30 in a check whose stated job
     on that field is validating a date. The same defect appeared in two modules three days
     apart, the second written after the first was fixed.
   disposition: fixed
   regression_test: >-
-    test_check_payload_rejects_impossible_dates and
-    test_apply_provenance_rejects_impossible_dates — behavioral, calling the validator rather
-    than grepping for `fromisoformat`. The check now requires BOTH the hyphenated shape and a
+    test_check_payload_rejects_impossible_dates — behavioral, calling the validator rather
+    than grepping for `fromisoformat`. The check requires BOTH the hyphenated shape and a
     successful parse: shape alone accepted 2026-99-99, and parse alone accepts the compact
     `20260815` from Python 3.11, which would put a second date spelling into a corpus whose
     schema declares `format: date`.
-
-- construct: the verification-clause boundary
-  canonical_owner: build/prose_provenance.py BOUND — and only as a PROPOSAL
-  duplicates:
-    - build/apply_provenance.py   # CLAUSE, removed
-  risk: >-
-    The sibling copy re-matched the clause at write time, so every boundary bug became a
-    silent corruption of prose rather than a visible wrong string. `[^.;]*` stopped inside
-    `huggingface.co`; adding a whitespace test then split `the U.S. AI Safety Institute
-    report` into two grammatical-looking halves, which is worse, because nothing looks wrong.
-
-    Deduplicating it would have missed the point. A heuristic good enough to PROPOSE a clause
-    for review is not good enough to rewrite prose unsupervised, so the applier now holds no
-    pattern at all: the boundary is confirmed once by a reviewer in the packet and substituted
-    literally. `BOUND` stays a heuristic and is labelled as one.
-  disposition: fixed
-  regression_test: >-
-    test_the_applier_holds_no_clause_pattern — AST, asserting the module neither imports nor
-    references `re`. A text search would have matched the docstring quoting the very fragments
-    it no longer uses, and would then have been weakened until it matched nothing.
-
-- construct: many-valued fact flattened to one value
-  canonical_owner: n/a
-  duplicates:
-    - build/apply_provenance.py   # score_sources
-  risk: >-
-    A URL is routinely cited on more than one axis with different `accessed` dates. `url ->
-    accessed` kept whichever instance came last in axis order, so the check answered "was the
-    LAST instance accessed on the claimed date?" while reporting an answer to "was one". Found
-    by the first real manifest: it refused two correct packets, and in the other direction
-    would have vouched for a URL whose date-aligned instance was not the one under review.
-  disposition: fixed
-  regression_test: test_a_url_cited_on_two_axes_is_checked_against_every_date_it_carries
-
-- construct: a composition nothing exercises
-  canonical_owner: n/a
-  duplicates:
-    - build/apply_provenance.py   # apply_one
-  risk: >-
-    `check` and `rewrite` were each tested alone and thoroughly. `apply_one`, which joins
-    them, referenced an undefined `CANONICAL` — so every real application would have raised
-    `NameError` after passing every check it was subjected to. The same shape as the rest of
-    this table, one level up: the parts were verified, the composition was not.
-  disposition: fixed
-  regression_test: test_an_unresolved_packet_of_any_state_applies — end to end, on a synthetic corpus
 
 - construct: openness class → bucket map
   canonical_owner: docs/openness-class-map.json
@@ -202,6 +155,11 @@ and tested, so it stays a decision instead of becoming drift.
 ## What was deliberately not done
 
 No new abstraction for the sake of deduplication. `build/vocabulary.py` holds two functions,
-both deriving from a file that already owns the fact. Constants that are genuinely local, or
+both deriving from a file that already owns the fact.
+
+Three entries were dropped in the 2026-08 cleanup along with the prose applier they described
+— a clause-boundary regex the applier should never have held, a `url -> accessed` map that
+flattened a many-valued fact, and an untested composition. Their lesson is the general one at
+the top of this file, and git carries the detail. Constants that are genuinely local, or
 that already have a working gate, were left where they are — noted above with the reason, so
 the next sweep does not rediscover them as findings.
