--- sources/provenance/agent_tools_protocols.yaml
@@ -1,176 +1,0 @@
-# Prose closeout for agent_tools_protocols. 32 of 32 ready.
-#
-# All 32 were canonical; 28 failed prose compliance. The ban detector flagged 15. Thirteen more
-# were found only by reading the complete prose of every record, which is now the second
-# category running where the detector's recall was well under half.
-#
-# What it missed, and why the pattern is worth writing down:
-#
-#   ADOPTION CLAIMS wearing description clothes - "Supported by more than 150 organizations"
-#   (agent2agent-protocol), "Powers search for over 18,000 customers in 150+ countries,
-#   including Stripe, Decathlon and Ubisoft" (algolia-ai-search), "used by several agent
-#   frameworks as their default scraping tool" (firecrawl), "Widely used in LangChain and agent
-#   frameworks" (serper), "used by thousands of production RAG systems" (pinecone).
-#
-#   RANKINGS phrased as category placement - "The most mature commercial web scraping platform"
-#   (apify), "The most-starred dedicated vector database on GitHub" (milvus), "The
-#   most-downloaded MCP package by a wide margin" (mcp-typescript-sdk), "The emerging universal
-#   standard for agent-tool communication" (model-context-protocol), "The default search
-#   provider in most agent tutorials" (tavily-search-api), "the most popular maintained one"
-#   (slack-mcp), "the canonical maintained one" (postgres-mcp), "the cheapest way to give agents
-#   Google-quality search results" (serper).
-#
-#   FUNDING in three spellings the money pattern did not cover - "YC-backed" (composio,
-#   browserbase, tavily-search-api), "Well-funded (YC-backed, $22M+ raised)" (exa-search-api),
-#   "Raised $500M+, valued at $9B+, the most well-funded company in the AI search space"
-#   (perplexity-sonar-api), "Raised $138M+ (Series B at $750M valuation)" (pinecone).
-#
-#   THE WORD `closed` on its own - "The enterprise closed alternative for search-augmented
-#   generation" (google-grounding-api), which the pattern only caught as `closed-source` or
-#   `closed counterpart`.
-#
-# ## One more stale-prose record, the seventh
-#
-# `browser-use`'s comments read "The adoption axis is flagged ... and carries no confirmation
-# date until that is settled". The axis carries 2026-08-14. Removed.
-#
-# ## Licensing questions kept, because they are about our reading
-#
-# This category has more of them than any other, because the MCP project is mid-transition and
-# its repositories disagree: `mcp-python-sdk` still carries the original file where
-# `mcp-typescript-sdk` and the servers monorepo have moved to the transition terms, and
-# `filesystem-mcp`'s own README and the monorepo now contradict each other. Each record says so
-# without stating which terms apply, because that is the axis's job.
-#
-# `mcp-inspector` and `markitdown` have no license file at all in the place a reader would look
-# - one declares in README and package.json, the other has a null PyPI field - and `yomo`'s two
-# branches disagree, which its comments record as an open question for the openness record.
-#
-- product: agent2agent-protocol
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the Linux Foundation's first-year release
-- product: algolia-ai-search
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Algolia AI Search product page
-- product: apify
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: apify
-- product: browser-use
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the pypistats API
-- product: browserbase
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: docs
-- product: cohere-rerank-api
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Cohere rerank overview, the 3
-- product: composio
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the Composio docs
-- product: docling
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the repository LICENSE and the pypistats API
-- product: exa-search-api
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the exa
-- product: fastmcp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the repository LICENSE and the pypistats API
-- product: filesystem-mcp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the server README and the monorepo LICENSE
-- product: firecrawl
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the self-hosting guide
-- product: github-mcp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the repository LICENSE
-- product: google-grounding-api
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Google Cloud grounding overview
-- product: jina-reader
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and jina
-- product: markitdown
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the repository LICENSE and the pypistats API
-- product: mcp-inspector
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the README and package
-- product: mcp-python-sdk
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the repository LICENSE and PyPI
-- product: mcp-typescript-sdk
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the repository LICENSE and package
-- product: milvus
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, milvus
-- product: model-context-protocol
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the spec repository LICENSE, the Wikipedia entry and Zuplo's one-year retrospective
-- product: notion-mcp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the repository LICENSE
-- product: perplexity-sonar-api
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: docs
-- product: pinecone
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Pinecone pricing page and docs
-- product: playwright-mcp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the repository LICENSE
-- product: postgres-mcp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the repository LICENSE
-- product: qdrant
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the qdrant
-- product: searxng
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, Docker Hub and the SearXNG search-API documentation
-- product: serper
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: serper
-- product: slack-mcp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the repository LICENSE
-- product: tavily-search-api
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: docs
-- product: yomo
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and Cargo
