--- sources/provenance/base_pretrained.yaml
@@ -1,160 +1,0 @@
-# Prose and provenance closeout for base_pretrained. 27 of 27 ready.
-#
-# Every record was already canonical; twenty still failed prose compliance. That is now the
-# second category in a row where the verification line was fine and the prose around it was not.
-#
-# ## Three records carried prose describing a score state the score had already left
-#
-#   ernie   "both the capability value and the adoption band need new sources" - both were
-#           re-sourced on 2026-08-14 to Baidu's own quarterly results.
-#   kimi    "the openness axis still records `weights - pending` and an assumed license, so it
-#           is behind the artifact" - the K3 LICENSE was read directly on 2026-08-14 and the
-#           axis moved with it.
-#   yi      "that record needs a new source" for capability - the axis carries a 2026-08-13 date
-#           against the model cards.
-#
-# With le-chat and vals-ai that makes five. The rule holds: a product's prose saying an axis is
-# unsettled is not evidence that it is, and the score note is what settles it.
-#
-# ## What came out
-#
-# Benchmark figures and leaderboard placements, as in finetuned_chat - `ernie`'s Arena rank and
-# score, `phi`'s "47th of 74 on the Artificial Analysis Intelligence Index", `kimi`'s ranking
-# against Claude Fable 5 and GPT-5.6 Sol, `falcon`'s "outperforms Qwen and Nemotron variants up
-# to 7x its size". Also download counts, license names throughout, and editorial framing that is
-# not description: `alia-40b`'s "the value proposition is permissive licensing, multilingual
-# breadth, and digital sovereignty", `inflection`'s "Notable as a case study in
-# foundation-model market concentration dynamics", `gpt-4o`'s "Listed as the historical
-# closed-source baseline that open-weights models were measured against".
-#
-# `minimax` and `qwen` each used ONE STRING for both prose fields - the seventh and eighth
-# instances found in the sweep.
-#
-# ## What stays, because for a base model it IS the record
-#
-# Parameter counts and active-versus-total for a mixture of experts, layer counts, context
-# windows, vocabulary size, training-token totals, language coverage, architecture, and who
-# trained it on what hardware. A new generation is a different record.
-#
-# The openness reasoning that distinguishes these entries also stays, because it is a judgment
-# call the score rests on rather than a restatement of it: `alia-40b`'s corpus is documented but
-# will not be distributed; `apertus` reaches the top tier on reconstruction scripts rather than
-# a released corpus; `lucie-7b`'s corpus is released but non-commercial, and the data dimension
-# asks whether it is released rather than on what terms; `rwkv` publishes an index of its
-# components but no corpus.
-#
-# ## One hold, unchanged
-#
-# `falcon.adoption` stays held since 2026-08-15: the declared checkpoint's real download count
-# lands a band below the recorded reported traction. Nothing in this pass touched it.
-#
-- product: alia-40b
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the langtech-bsc/alia repository
-- product: apertus
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the v1
-- product: deepseek
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the DeepSeek-V4-Pro HF model card and the deepseek-ai/DeepSeek-V3 repository
-- product: ernie
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the ERNIE 5
-- product: falcon
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Falcon 3 release blog and the Falcon3 model cards
-- product: gemma
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the gemma-4-31B-it model card and the Gemma releases page
-- product: glm
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GLM-5
-- product: gpt-4o
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: OpenAI's GPT-4o model documentation
-- product: inflection
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Intel newsroom release and the Inflection AI Wikipedia entry
-- product: inkling
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Inkling model card and Thinking Machines' launch post
-- product: internlm
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the InternLM repository
-- product: kimi
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the Kimi-K3 model card and its LICENSE file
-- product: llama
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Llama 4 Scout HF model card and the meta-llama organization page
-- product: lucie-7b
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card, the dataset card and the Lucie-Training repository
-- product: luciole
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the 23B base and instruct cards, the dataset card and the Luciole-Training repository
-- product: marin
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Marin announcement post, the marin-8b-base HF model card and the marin-community/marin
-    repository
-- product: mimo-pro
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card, the product page and the MiMo repository
-- product: minimax
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the M3 and M2
-- product: mistral-large
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Mistral 3 launch post and the Mistral-Large-3-675B-Base-2512 HF model card
-- product: olmo
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Ai2 OLMo 3 blog, the Olmo-3-1025-7B model card and the allenai/OLMo repository
-- product: phi
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the microsoft/phi-4 model card and the Artificial Analysis model page
-- product: pythia
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Pythia paper, the Pile dataset entry, the EleutherAI/pythia repository and the pythia-12b
-    model card
-- product: qwen
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Qwen3
-- product: reka-core
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Reka Core launch page
-- product: rwkv
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the component index, the dataset repository, RWKV-LM and the model repository
-- product: smollm
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the SmolLM3 release post, the SmolLM3-3B model card and the huggingface/smollm repository
-- product: yi
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Yi-1
