--- sources/provenance/benchmark_eval_data.yaml
@@ -1,145 +1,0 @@
-# Prose and provenance closeout for benchmark_eval_data. 27 of 27 ready.
-#
-# This category held most of the corpus's `named_noncanonical` records - the ones whose lines
-# read `Verified live <date> on HF (285,003 downloads in the trailing 30 days)`. A host and a
-# download count is not a document, which is why they were never canonical: the line named where
-# somebody looked, not what settled it. Each now names the dataset card.
-#
-# Four records used ONE STRING for both prose fields - gpqa, mmmu, truthfulqa and, in the earlier
-# categories, lovable and github-copilot. A description repeated as a footnote is the failure
-# mode `product-copy.md`'s division of labor exists to prevent, and it is invisible to every gate.
-#
-# ## What a benchmark record may not carry
-#
-# Model scores. `livecodebench` listed four frontier models with their percentages;
-# `swe-bench-verified` carried an April 2026 audit finding no source in the record establishes;
-# `evalplus` quoted o1's pass@1. How a MODEL scores on a benchmark belongs to the model, and how
-# the benchmark itself scores belongs to its capability axis. Also removed: download counts,
-# GitHub stars, and rankings ("the gold-standard benchmark", "the standard benchmark for
-# multilingual code models", "the trusted dynamic complement to HumanEval/MBPP").
-#
-# Kept: sizes, splits, task counts, publication venue and construction method. Those are what the
-# dataset IS, and a new version would be a different record.
-#
-# ## Three access findings that are not evidence holds
-#
-# `math`: the canonical Hugging Face repository is access-disabled under a DMCA takedown, so the
-# data is no longer redistributable from the registry and its trailing download count is zero.
-# The axis keeps its date - the page is the evidence and it was read - and the prose says so.
-#
-# `swe-bench-verified` and `livecodebench`: the dataset cards declare no usable license (none at
-# all, and a bare `cc` naming no variant). Recorded as ambiguities for the openness axis to
-# resolve, not as verdicts in prose. `multipl-e` carries the opposite problem - a license clause
-# forbidding use of the contents as training data - noted the same way.
-#
-# `gaia` and `humanitys-last-exam` hold their answers server-side; `mmmu` used to and published
-# them on 2026-02-12. Those are facts about how the benchmark is scored, so they stay.
-#
-- product: ai2-arc
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the allenai/ai2_arc dataset card on Hugging Face
-- product: anthropic-internal-coding-evals
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: Anthropic's Claude Opus 4
-- product: apps
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the codeparrot/apps dataset card on Hugging Face
-- product: codecontests
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the deepmind/code_contests dataset card on Hugging Face
-- product: compar-ia-datasets
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the ministere-culture/comparia-votes dataset card on Hugging Face
-- product: cruxeval
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the facebookresearch/cruxeval repository
-- product: evalplus
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the evalplus/humanevalplus dataset card and the evalplus/evalplus repository
-- product: gaia
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GAIA dataset card on Hugging Face and the GAIA paper abstract
-- product: gpqa
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Idavidrein/gpqa dataset card on Hugging Face and the GPQA paper abstract
-- product: gsm8k
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the openai/gsm8k dataset card on Hugging Face
-- product: hellaswag
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Rowan/hellaswag dataset card on Hugging Face
-- product: humaneval
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the openai/openai_humaneval dataset card on Hugging Face
-- product: humanitys-last-exam
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the cais/hle dataset card on Hugging Face and lastexam
-- product: ifeval
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the google/IFEval dataset card on Hugging Face
-- product: livebench
-  resolved_state: canonical
-  verified: '2026-08-12'
-  document: the repository DATASHEET
-- product: livecodebench
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the livecodebench/code_generation_lite card, its raw README and the harness repository
-- product: math
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the hendrycks/competition_math dataset page
-- product: mbpp
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the google-research-datasets/mbpp dataset card on Hugging Face
-- product: mmlu
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the cais/mmlu dataset card on Hugging Face
-- product: mmlu-pro
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the TIGER-Lab/MMLU-Pro dataset card on Hugging Face and the MMLU-Pro paper abstract
-- product: mmmu
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the MMMU/MMMU dataset card on Hugging Face and the MMMU paper abstract
-- product: mt-bench
-  resolved_state: canonical
-  verified: '2026-08-12'
-  document: the FastChat llm_judge README, the FastChat LICENSE and the paper
-- product: multipl-e
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the nuprl/MultiPL-E dataset card and the repository LICENSE
-- product: openai-internal-evals
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenAI Preparedness Framework v2 PDF
-- product: scale-seal-leaderboard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: Scale AI's leaderboard announcement
-- product: swe-bench-verified
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the SWE-bench_Verified card, its raw README and the SWE-bench repository LICENSE
-- product: truthfulqa
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the truthfulqa/truthful_qa dataset card on Hugging Face and the paper abstract
