--- sources/provenance/dataset_processing_tools.yaml
@@ -1,101 +1,0 @@
-# Prose closeout for dataset_processing_tools. 21 of 21 ready. Six failed; the ledger flagged 4.
-#
-# `datology-ai` carried the most banned content of any record in the sweep: "a proprietary
-# curation platform", "no source-available product", "~$57.5M raised (seed plus a $46M Series A
-# in May 2024)", and a named angel list. What survives is the founders and the fact that no
-# frontier-lab customer is disclosed, which is why the adoption axis has nothing to band.
-#
-# `snorkel-flow` carried a 120-word comments field, and it was long because it records something
-# worth keeping: a 2026-08-13 pass read only the marketing site, where the old path now
-# redirects and the copy sells data services, and concluded the platform had been withdrawn. The
-# documentation site showed it had only been renamed. THE LESSON IS THE POINT - a vendor
-# reorganising its marketing is not evidence a product is gone - and it now fits in 48 words.
-#
-# `txt360-pipeline` keeps its licensing ambiguity because it explains the treatment: the README
-# badge asserts a license the repository ships no file for, and the GitHub API reports none,
-# which is why openness abstains rather than scoring it.
-#
-- product: bigcode-dataset
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: ccnet
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: data-juicer
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: datamap-rs
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: datatrove
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: datology-ai
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: datologyai
-- product: dclm
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the repository README
-- product: decon
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: distilabel
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: dolma-toolkit
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: duplodocus
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: marin-framework
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: nemo-curator
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: nemo-data-designer
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the NVIDIA NeMo product page and the DataDesigner README
-- product: olmocr
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: redpajama-data
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: smallpond
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: snorkel-flow
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: docs
-- product: the-pile-pipeline
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
-- product: txt360-pipeline
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the repository API
-- product: ungoliant
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GitHub README and the LICENSE body
