--- sources/provenance/deployment.yaml
@@ -1,125 +1,0 @@
-# Prose closeout for deployment. 27 of 27 ready. Five failed; the ledger flagged 2.
-#
-# Three failed on LENGTH alone, which no ban pattern catches: aws-lambda at 80 description
-# words, firecracker at 86, ray at 90, against the guide's 35-70 band. Long descriptions are
-# where the other violations hide, and reading for length is what surfaced them.
-#
-# `spaces` carried a 93-word comments field recording a real correction: its declared GitHub
-# artifact was gradio-app/gradio, removed 2026-08-14 because that repository is an app SDK that
-# runs anywhere, is already `gradio`'s declared artifact, and would have banded Spaces on
-# Gradio's stars. Spaces publishes no countable channel of its own, so nothing replaced it. The
-# correction is kept; the retelling is shorter.
-#
-# `aws-lambda` and `google-cloud-run` both stated the openness verdict outright. The
-# Lambda-to-Firecracker relationship stays, because one is the substrate of the other and both
-# are records on this map.
-#
-- product: agent-infra-sandbox
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the agent-infra/sandbox README and the GitHub repository metadata
-- product: aws-lambda
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the AWS Lambda FAQ
-- product: beta9
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Beta9 README and the ComputeSDK sandbox benchmark leaderboard
-- product: blaxel-sandbox
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: blaxel
-- product: cloudflare-sandboxes
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: Cloudflare's "Sandboxing AI agents, 100x faster" post and the cloudflare/sandbox-sdk repository
-- product: coder-oss
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the coder/coder README and the GitHub repository metadata
-- product: daytona-sandbox
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: daytona
-- product: e2b-sandbox
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the E2B site and the ComputeSDK sandbox benchmark leaderboard
-- product: firecracker
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Firecracker README and the GitHub repository API
-- product: fly-sprites
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: fly
-- product: google-cloud-run
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: cloud
-- product: gvisor
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the gVisor README, the LICENSE body, and the gvisor
-- product: kata-containers
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Kata Containers README, docs/hypervisors
-- product: modal-sandboxes
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Modal Sandboxes guide and the Modal pricing page
-- product: nono
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the nono README, the LICENSE body, and nono
-- product: northflank-sandboxes
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: northflank
-- product: ollama
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the repository README, the LICENSE body, ollama
-- product: openai-code-interpreter
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenAI Code Interpreter developer documentation
-- product: openpcc
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenPCC and Confident-Compute READMEs, the LICENSE body, and the Cloud Security Alliance
-    launch write-up
-- product: opensandbox
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenSandbox README, the secure container runtime guide, and open-sandbox
-- product: ray
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Anyscale Ray documentation and the PyPI download API
-- product: replit-agent-code-execution-api
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the replit/replit-code-exec repository and its README
-- product: sandbox-runtime
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the anthropic-experimental/sandbox-runtime README and the GitHub repository metadata
-- product: spaces
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Spaces documentation and pricing page
-- product: tensorlake-sandbox
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: tensorlake
-- product: vercel-sandbox
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the Vercel Sandbox documentation and its pricing and quotas page
-- product: webcontainers
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: webcontainers
