--- sources/provenance/edge_hardware.yaml
@@ -1,128 +1,0 @@
-# Prose and provenance closeout for edge_hardware. 20 of 20 ready.
-#
-# This category closes the provenance sweep: with it, all 472 products are canonical and the
-# census reports zero unresolved records.
-#
-# ## The last four `substitute sources` records
-#
-# `hailo-8`, `hailo-10h`, `nxp-imx-8m-plus` and `qualcomm-dragonwing-rb3-gen-2` were the four
-# records pulled out of the class-A rewording in PR #282, because removing `live` from
-# "Verified live <date> via substitute sources" would have promoted a method into canonical
-# form. Each now names the documents actually read, and each says which vendor host refused:
-# hailo.ai answers 403, nxp.com answers 404 to its own root, qualcomm.com serves a
-# client-rendered shell. The substitutes are first-party where possible - NXP's own datasheet
-# mirrored by DigiKey, Qualcomm's own product brief on docs.qualcomm.com - and named as such.
-#
-# `qualcomm-dragonwing-rb3-gen-2` lost two clauses that no substitute could re-derive: a
-# QCS5430 variant and Foundries.io support. Neither appears in any reachable source.
-#
-# ## What came out
-#
-# Retail prices on seven records (EUR 47-65, $72, $70, $149, $249, $69-$115), which move; the
-# openness verdict, which for hardware is the whole substance of the axis; and framing that is
-# not description - "a low-cost on-ramp to Qualcomm edge-AI silicon for the maker community",
-# "the cheapest on-ramp into the CUDA/Jetson generative-AI ecosystem", "it matters as an
-# ultra-low-cost RISC-V edge-AI vision module with strong open tooling", "the official
-# first-party way to add meaningful NPU compute to a Pi 5".
-#
-# What stays is the part specification: TOPS at a named precision, core counts and clocks, GPU
-# and NPU blocks, memory configurations, form factor, process node, interfaces. For a board or
-# a chip that IS the record, and a new revision is a different record.
-#
-# Where the openness reading is genuinely load-bearing, `comments` names the tension without
-# resolving it - published board files and toolchain against silicon that needs vendor firmware
-# blobs; a released runtime against a registration-gated compiler; a public datasheet against
-# an SDK shipped as binaries.
-#
-# ## One record whose platform is ending
-#
-# `google-coral-dev-board`: the Edge TPU libraries are archived read-only, coral.ai's product
-# page redirects to one naming no hardware, and both variants are discontinued at retail. Its
-# adoption level and retail component were both resolved on 2026-08-14; the prose had still
-# said they were "escalated for a ruling", which is the sixth stale-prose instance in the sweep.
-#
-- product: arduino-uno-q
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the Arduino product page, the App Lab and App Bricks repositories and the UNO Q user manual
-- product: axelera-metis-aipu
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Axelera product site and the voyager-sdk repository
-- product: beagley-ai
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OSHWA certification record, the beagley-ai repository, the design documentation and
-    the TI product page
-- product: google-coral-dev-board
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the datasheet, the electricals repository and the product page
-- product: hailo-10h
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: those two listings and the GenAI model-zoo repository
-- product: hailo-8
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: those listings and the HailoRT README
-- product: khadas-edge2
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Khadas schematics archive, the fenix toolchain repository, the Edge2 product page
-    and the rkbin repository
-- product: memryx-mx3
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the M
-- product: nvidia-jetson-agx-orin
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Jetson Orin product page and the Jetson Linux developer page
-- product: nvidia-jetson-orin-nano-super-developer-kit
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the developer-kit page, the Jetson Orin product page and the Jetson Linux developer page
-- product: nvidia-jetson-orin-nx
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Jetson Orin product page and the Jetson Linux developer page
-- product: nxp-imx-8m-plus
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: those three
-- product: orange-pi-5
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Orange Pi 5 product and support pages, the orangepi-build repository and the rkbin
-    repository
-- product: qualcomm-dragonwing-rb3-gen-2
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: those two
-- product: radxa-rock-5b
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Radxa download and introduction pages, the radxa-repo/bsp repository and the rkbin
-    repository
-- product: raspberry-pi-5
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the product brief, the mechanical drawing and the firmware README
-- product: raspberry-pi-ai-hat-plus
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: those two datasheets and the HailoRT README
-- product: rockchip-rk3588
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Rockchip brief datasheet, the product page, the rknn-toolkit2 repository and the rkbin
-    repository
-- product: sipeed-maixcam
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Sipeed MaixCam wiki page and the MaixPy repository
-- product: ti-am67a
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the TI product page, the edgeai-tidl-tools repository and the BeagleY-AI board page
