--- sources/provenance/evaluation_code.yaml
@@ -1,150 +1,0 @@
-# Prose and provenance closeout for evaluation_code. 26 of 26 ready.
-#
-# All 26 were unresolved except `openai-evals-api-dashboard` and `scale-seal-leaderboard`, both
-# of which were already canonical and still carried curator rationale and pricing. No axis was
-# held, no date moved, no source refetched.
-#
-# ## One stale prose claim, caught the way le-chat was
-#
-# `vals-ai`'s comments read "the openness score has not yet been reconciled with that" - meaning
-# with the two repositories Vals had since released. The score file says the opposite: the note
-# records the 2026-08-14 re-read that applied exactly that correction and moved the axis from
-# 1/closed to 2/source_available. The prose was describing a state the score had already left.
-#
-# That is the same shape as le-chat in ui_api, and the rule it produced held here: read the
-# score note before believing what a product says about its own axes.
-#
-# ## What came out
-#
-# Curator rationale was heavier in this category than any other - eleven records opened a clause
-# with "Picked because", "Picked when", "Picked over" or "Picked by". `product-copy.md` calls
-# that our selection judgment rather than a description, and it is where the banned content had
-# collected: star counts, funding ("Raised $20M", "Raised a seed round via AIGrant"), pricing
-# ("$10-$20 per 1,000 evaluator API calls"), and rankings ("the de-facto-standard model-eval
-# harness", "the canonical reference harness", "the authoritative scoreboard").
-#
-# Model scores came out for the same reason they did in benchmark_eval_data: `osworld` listed
-# three frontier models with their percentages, `swe-bench` carried an April 2026 audit finding
-# no source in the record establishes.
-#
-# ## Lifecycle facts kept, because they are what these records are for
-#
-# Half this category is deprecated, archived or sunsetting, and each says so plainly:
-# `openai-evals-api-dashboard` (read-only 2026-10-31, shutdown 2026-11-30), `simple-evals`
-# (deprecated July 2025), `helm` (maintenance mode since 2026-06-01), `open-llm-leaderboard`
-# (archived March 2025), `bigcodebench` (repository archived January 2026), `openai-evals` (no
-# push since April 2026), `ragas` (no push since February 2026), `chatbot-arena` (FastChat and
-# Arena-Hard-Auto both stale since 2025). A statement of absence is durable by the guide's own
-# test, and for an eval harness it is the single most useful thing a reader can know.
-#
-# ## Harness-versus-dataset boundaries recorded
-#
-# `bigcodebench`, `osworld` and `swe-bench` each ship both a task set and a runner. Each record
-# now says which half it scores and points at benchmark_eval_data for the other.
-#
-- product: alpacaeval
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the tatsu-lab/alpaca_eval repository, its license endpoint and its README
-- product: amazon-bedrock-evaluations
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Bedrock Evaluations product page and its general-availability announcement
-- product: artificial-analysis-intelligence-index
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Artificial Analysis benchmarking methodology page and the Stirrup license endpoint
-- product: bigcodebench
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the bigcodebench repository, its license endpoint and its README
-- product: chatbot-arena
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: arena
-- product: compar-ia
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the betagouv/ComparIA repository README and its LICENSE
-- product: deepeval
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the confident-ai/deepeval repository, its license endpoint and its README
-- product: epoch-ai-benchmarks
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Epoch benchmarks dashboard, the FrontierMath page and the organization's repository
-    listing
-- product: helm
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the stanford-crfm/helm repository, its license endpoint and its README
-- product: inspect-ai
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the UKGovernmentBEIS/inspect_ai repository, its license endpoint and the Inspect models
-    documentation
-- product: lighteval
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the huggingface/lighteval repository, its license endpoint and its README
-- product: lm-evaluation-harness
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the EleutherAI/lm-evaluation-harness repository, its license endpoint and its README
-- product: open-llm-leaderboard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Space README, the Space API and the leaderboard archive page
-- product: openai-evals
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the openai/evals README, its LICENSE
-- product: openai-evals-api-dashboard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenAI Evals guide
-- product: opencompass
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the open-compass/opencompass repository and its license endpoint
-- product: osworld
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the xlang-ai/OSWorld repository, its license endpoint and the project site
-- product: patronus-evaluation-platform
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Patronus documentation and the patronus-py repository
-- product: promptfoo
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the promptfoo repository, its license endpoint and its README
-- product: ragas
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the vibrantlabsai/ragas repository, its license endpoint and its README
-- product: scale-evaluation
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Scale GenAI platform page, the leaderboard announcement and the SEAL Showdown post
-- product: simple-evals
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the openai/simple-evals repository, its license endpoint and its README
-- product: simpleaudit
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the simpleaudit repository README and its LICENSE
-- product: swe-bench
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the SWE-bench repository, its license endpoint and its README
-- product: vals-ai
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the Vals AI about page and the two repositories' metadata
-- product: vertex-ai-gen-ai-evaluation-service
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Vertex AI evaluation overview documentation
