--- sources/provenance/finetuned_chat.yaml
@@ -1,199 +1,0 @@
-# Prose and provenance closeout for finetuned_chat. 39 of 39 ready.
-#
-# Every record in this category was ALREADY CANONICAL. Twenty-two still failed prose compliance,
-# which is the clearest demonstration yet that classifier state cannot scope this work: the
-# verification line named a document in all 39 cases and said nothing about what the prose
-# around it claimed.
-#
-# ## What a model record may not carry
-#
-# Benchmark scores, and they were everywhere. `claude-opus` carried five SWE-bench percentages
-# and an LMArena Elo; `gemini-flash` carried an Elo, a LiveCodeBench figure and three more;
-# `devstral`, `deepseek-instruct`, `doubao-seed` and `muse-spark` each carried their own. A
-# model's score on a benchmark is what the capability axis exists to hold, with a source and a
-# date behind it. In prose it is a number with no owner that a single re-run invalidates.
-#
-# Also removed: API pricing (`claude-fable` $10/$50 per Mtok, `doubao-seed` $0.47/$2.37),
-# rankings ("the first non-Anthropic/Google/OpenAI model to break the top 5", "strongest
-# fully-open model by parameter scale", "one of the most-pulled 7B checkpoints on the Hub"),
-# adoption figures that belong to a different product (`doubao-seed` carried the Doubao APP's
-# DAU and MAU), and license names throughout.
-#
-# Two claims came out because no source in the record establishes them: `claude-opus`'s "leaked
-# 'Claude Mythos Preview' tops SWE-bench at 93.9%", and `codestral`'s "distributed to thousands
-# of developers ... reported usage trails Anthropic and OpenAI coding models".
-#
-# ## What stays, because for a model it is identity
-#
-# Parameter counts, active-versus-total for a mixture of experts, context windows, output
-# limits, release dates, base model, training-token counts and post-training method. A new
-# release is a different record, so none of it is volatile in the guide's sense.
-#
-# ## Identity notes this category turns on
-#
-# Half these records are compound or consolidated, and each says which SKU it scores:
-# `deepseek-instruct` scores V4-Flash while V4-Pro is a frozen base_pretrained anchor;
-# `claude-opus` scores the 4.x line while 4.7 is the frozen anchor; `hermes` follows the 36B
-# Apache build across four differently-licensed bases; `devstral` was rescored onto the
-# governing 2512 release; `amazon-nova`, `gemini-flash` and `claude-opus` were consolidated on
-# 2026-07-29; `starcoder2` scores the instruct SKU while its declared artifact is the base repo.
-#
-# `starcoder2` also used ONE STRING for both prose fields, the sixth instance found so far.
-#
-- product: amazon-nova
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the AWS Nova user guide and models page
-- product: claude-fable
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Fable 5 launch post, the Claude platform documentation and the Claude Fable product
-    page
-- product: claude-haiku
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Claude Haiku product page
-- product: claude-opus
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Opus 4
-- product: claude-sonnet
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Sonnet 4
-- product: codegemma
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and Hugging Face's CodeGemma release post
-- product: codellama
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the arXiv abstract
-- product: codestral
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Codestral and Codestral 25
-- product: command-r
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the Command A+ model card and the CohereLabs org listing
-- product: deepseek-coder
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the model-license body
-- product: deepseek-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the DeepSeek-V4-Flash model card
-- product: deepseek-r1
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: devstral
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: both model cards
-- product: doubao-seed
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: TechNode's coverage of the launch
-- product: gemini-flash
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Gemini 3
-- product: gemini-pro
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the DeepMind model index
-- product: gpt-4-1
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: OpenAI's model docs, TechCrunch and InfoQ
-- product: gpt-5
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the GPT-5 Wikipedia article and TechCrunch
-- product: granite-code-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card, the granite-code-models repository and IBM's release blog
-- product: grok
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the xAI release notes, the xAI model reference and the Artificial Analysis model page
-- product: hermes
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model cards and the technical report
-- product: jamba-large
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card, the Jamba Open Model License body and the AI21 release blog
-- product: k2
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the arXiv abstract
-- product: llama-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: mistral-7b-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card
-- product: muse-spark
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: Meta's newsroom post
-- product: nemotron
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Nemotron 3 page, NVIDIA's developer blog and the Super technical report
-- product: o-series
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenAI o3/o4-mini system card and the Wikipedia article
-- product: olmo-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the OLMo 3 blog
-- product: olmoe-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the OLMoE release post
-- product: ornith
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model cards
-- product: phi-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: qwen-coder
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: qwen-instruct
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: starcoder2
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: both model cards
-- product: tinyllama-chat
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card
-- product: tulu
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model cards, the Tulu 3 blog and the open-instruct repository
-- product: vibethinker
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: zephyr
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the model card and the alignment-handbook repository
