--- sources/provenance/finetuning_code.yaml
@@ -1,133 +1,0 @@
-# Prose closeout for finetuning_code. 27 of 27 ready. Four records failed, as flagged.
-#
-# ## The distinction this category forced
-#
-# Most records here say they fine-tune "open-weight models". That is NOT this product's own
-# openness - it is the class of input the service accepts, which is a capability fact and the
-# single most useful thing in the entry. It stays on anyscale-fine-tuning,
-# azure-openai-fine-tuning, fireworks-fine-tuning, tinker and together-fine-tuning.
-#
-# What came out is the openness verdict applied to a NAMED ARTIFACT: nemo-rl's "NVIDIA's open
-# source library" about itself, predibase's "the multi-LoRA inference engine its founders
-# publish as open source", mistral-fine-tuning-api's "A separate open source SDK", and
-# openai-fine-tuning-api's "OpenAI's closed models".
-#
-# `openai-fine-tuning-api` keeps its wind-down in full - new sign-ups stopped, job creation ends
-# January 2027, fine-tuned models serve until their base models are deprecated - because "closed
-# to new users" is a lifecycle fact rather than an openness verdict, and it is what a reader
-# needs first.
-#
-# `verl.capability` remains held since 2026-08-09 and was not touched.
-#
-- product: amazon-bedrock-custom-models-fine-tuning
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the AWS Bedrock model-customization documentation and AWS's Claude 3 Haiku fine-tuning
-    GA announcement
-- product: anyscale-fine-tuning
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Anyscale fine-tuning documentation and the Terms of Service
-- product: atropos
-  resolved_state: canonical
-  verified: '2026-08-08'
-  document: the GitHub repository and the PyPI record for atroposlib
-- product: axolotl
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the README, the LICENSE body, and axolotl
-- product: azure-openai-fine-tuning
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Microsoft Foundry fine-tuning documentation, the Azure OpenAI data-privacy page, and
-    Microsoft's Product Terms
-- product: fireworks-fine-tuning
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Fireworks fine-tuning, model catalog, LoRA-deployment, and LoRA-performance documentation
-- product: lamini
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: docs
-- product: litgpt
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: llama-factory
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: megatron-lm
-  resolved_state: canonical
-  verified: '2026-08-08'
-  document: the NVIDIA/Megatron-LM README and the Megatron Core developer documentation
-- product: mistral-fine-tuning-api
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: Mistral's deprecated fine-tuning guide, the Studio product page, and the API pricing page
-- product: mosaic-ai-model-training
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Databricks Foundation Model Fine-tuning docs page and the Databricks Model Training
-    product page
-- product: nemo-rl
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, and PyPI
-- product: openai-fine-tuning-api
-  resolved_state: canonical
-  verified: '2026-08-08'
-  document: the OpenAI model optimization guide, the API deprecations page, and the OpenAI Services
-    Agreement
-- product: openpipe
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the ART documentation
-- product: openrlhf
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: peft
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, and Hugging Face's pricing page
-- product: predibase
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: web
-- product: snowflake-cortex-fine-tuning
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Snowflake Cortex Fine-tuning documentation
-- product: swift
-  resolved_state: canonical
-  verified: '2026-08-08'
-  document: the ms-swift README, the PyPI project page, and the LICENSE body
-- product: tinker
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Tinker product page, the Service Terms of Use, and Thinking Machines' news archive
-- product: together-fine-tuning
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Together AI fine-tuning docs, the fine-tuning product page, and the Terms of Service
-- product: torchtune
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: trl
-  resolved_state: canonical
-  verified: '2026-08-08'
-  document: the TRL documentation index and the repository README
-- product: unsloth
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: verl
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, the README, and PyPI
-- product: vertex-ai-tuning
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Gemini tuning documentation and the Google Cloud Platform Terms of Service
