--- sources/provenance/inference_code.yaml
@@ -1,99 +1,0 @@
-# Prose closeout for inference_code. 20 of 20 ready. One failed, as flagged.
-#
-# `tensorrt-llm` named the third-party licenses bundled in its repository; the note now says a
-# compound exists and which half the axis follows, without restating either.
-#
-# ## The one detector false positive in the whole sweep
-#
-# `text-generation-inference` reports "Its final release was v3.3.7 (2025-12-19) and the
-# repository was archived 2026-03-21". The ban detector reads that as a current version number.
-# It is not: `product-copy.md` names this exact case as durable - "There will be no future
-# release ... durable precisely because the project stopped" - and gives an almost identical
-# example. The prose was left alone and the detector is wrong here, which is recorded rather
-# than worked around.
-#
-# `qualcomm-ai-engine-direct.openness` remains held since 2026-08-09 and was not touched.
-#
-- product: apple-core-ml-runtime
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: Apple's Core ML documentation, the MLModel availability list, and the coremltools optimization
-    guide
-- product: aws-neuron
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the AWS Neuron documentation, the aws-neuron-sdk and aws-neuron-driver GitHub repos, and
-    the AWS Neuron product page
-- product: cerebras-inference
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Cerebras inference page, the Cerebras chip page, and the Cerebras Inference Terms
-    and Conditions
-- product: ds4
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: fireworks-inference
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Fireworks homepage, team page, and FireAttention V3 blog post
-- product: google-cloud-tpu-inference
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the Google Cloud AI Hypercomputer inference blog post and the TPU7x (Ironwood) documentation
-- product: groq-inference
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the GroqDocs models and overview pages and Groq's LPU architecture blog
-- product: llama-cpp
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, and the Hugging Face GGUF documentation
-- product: llamafile
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, and the README
-- product: lmdeploy
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, and PyPI download statistics
-- product: localai
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: max
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the Modular Community License page
-- product: onnx-runtime
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, and PyPI download statistics
-- product: qualcomm-ai-engine-direct
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the ExecuTorch Qualcomm backend docs, and the LiteRT NPU documentation
-- product: sambanova-cloud
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: the SN40L architecture blog and the SambaCloud Terms of Service
-- product: sglang
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: tensorrt-llm
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
-- product: text-generation-inference
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub, the LICENSE body, and the releases page
-- product: together-inference-engine
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: together
-- product: vllm
-  resolved_state: canonical
-  verified: '2026-08-09'
-  document: GitHub and the LICENSE body
