--- sources/provenance/ml_frameworks.yaml
@@ -1,111 +1,0 @@
-# Prose closeout for ml_frameworks. 22 of 22 ready.
-#
-# All 22 were canonical; 12 failed prose compliance where the ledger flagged 6.
-#
-# This category's records are the most established software on the map, and that is exactly
-# what put the violations in: nine of the twelve were a claim about the product's STANDING
-# rather than about the product. "It is one of the two most widely used deep learning
-# frameworks and underpins most modern research and production model training" (pytorch),
-# "one of the two most widely used" (tensorflow), "it has become the default weight format for
-# models on the Hugging Face Hub" (safetensors), "widely used as the kernel backend" (triton),
-# "widely used to feed training and evaluation pipelines" (hf-datasets), "widely used as the
-# distributed-training backbone" (accelerate), "widely used for high-performance numerical
-# computing" (jax), "one of the most established open frameworks" (pysyft), "underpinning much
-# of the modern open source ML and LLM ecosystem" (transformers).
-#
-# Each was replaced with the same fact stated as a relationship rather than a rank - "the
-# framework most of the model records on this map are trained and served with", "other Hugging
-# Face libraries use it as their distributed-training backbone", "serves as the kernel backend
-# for PyTorch's inductor". That is checkable; "most widely used" is not.
-#
-# The remaining three opened with the openness verdict as the category noun: flax, gradio,
-# openvino.
-#
-- product: accelerate
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: deepspeed
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: diffusers
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: feluda
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the tattle-made/feluda repository and the GitHub API
-- product: flax
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the GitHub license API
-- product: ggml
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: gradio
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: hf-datasets
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: huggingface-hub
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: jax
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the GitHub license API
-- product: keras
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the GitHub license API
-- product: onnx
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: openvino
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: optimum
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: pysyft
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenMined/PySyft repository, its README and the PyPI project page
-- product: pytorch
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: safetensors
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: sentencepiece
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: tensorflow
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the GitHub license API
-- product: tokenizers
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: transformers
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
-- product: triton
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and the LICENSE body
