--- sources/provenance/safeguards.yaml
@@ -1,154 +1,0 @@
-# Prose closeout for safeguards. 26 of 26 ready.
-#
-# All 26 were canonical; 18 failed prose compliance. The detector flagged 10 and missed 8.
-#
-# ## The failure shape specific to this category
-#
-# A guardrail product's deployment model is the first thing a reader wants to know, so eight
-# records ended their description with the verdict as a two-word sentence - "Closed, managed
-# service." (azure-content-safety, bedrock-guardrails), "Closed/SaaS" (lakera-guard), "Closed,
-# API-only; no weights or self-hosting" (openai-moderation, perspective-api) - or opened with
-# it: "An open-source red-teaming framework" (deepteam), "An open security toolkit" (llm-guard),
-# "An open framework" (guardrails-ai), "An open guardrails project" (openguardrails).
-#
-# That is the openness axis stated twice, once with evidence and once without. What replaced it
-# is the deployment mechanism, which is architecture a reader can act on: "run entirely as an
-# Azure service", "runs as a Python library on infrastructure you operate, or deployed as an
-# API", "deploys as a Python library, a CLI, a Docker image or a standalone server".
-#
-# ## What the detector missed
-#
-#   llama-guard        "a widely deployed open guardrail model" and "Widely used as the
-#                      moderation layer in front of open chat and agent stacks" - two adoption
-#                      claims in a description
-#   wildguard          "benchmarked above GPT-4 on several moderation tasks" - capability
-#   openguardrails     "The first project to open-source both a large-scale safety LLM and a
-#                      production guardrails platform"
-#   osprey             "processing over 50M events daily in production", plus a named
-#                      deployment list
-#   aegis-guard        "so it is not fully open-weight"
-#   gpt-oss-safeguard  "under Apache 2.0"
-#   garak              "the nmap of LLM security testing"
-#   harmbench          "run entirely from the public repo"
-#
-# ## Gating notes KEPT, because they explain treatment rather than making the verdict
-#
-# Four model records say their weights need an access request - aegis-guard, llama-guard,
-# llama-prompt-guard, shieldgemma - and each says it as the reason openness was read from the
-# card rather than from a download. That is a statement about how the evidence was obtained. It
-# is not the same as asserting a tier, and it stays.
-#
-# The same test keeps the missing-training-data notes on granite-guardian, qwen3guard,
-# shieldgemma and gpt-oss-safeguard, and the adapter-versus-full-model note on zentropi-cope.
-#
-# ## One lifecycle fact worth more than any score
-#
-# `perspective-api` is sunsetting: the service runs until 31 December 2026, new usage and quota
-# requests closed in February 2026, and Jigsaw is offering no migration support. That is in the
-# description and the comments both, because a reader choosing a toxicity API this year needs it
-# before anything else on the record.
-#
-- product: aegis-guard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: azure-content-safety
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Azure product page
-- product: bedrock-guardrails
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the AWS product page
-- product: coop
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and ROOST's Coop 1
-- product: deepteam
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: garak
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: giskard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: gpt-oss-safeguard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: granite-guardian
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: guardrails-ai
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: harmbench
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: lakera-guard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Lakera site
-- product: llama-guard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: llama-prompt-guard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: llamafirewall
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the LlamaFirewall README and the PurpleLlama repo
-- product: llm-guard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: nemo-guardrails
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: openai-moderation
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the OpenAI moderation guide
-- product: openguardrails
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub, the HF model card and the arXiv abstract
-- product: osprey
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub and ROOST's Osprey announcement
-- product: perspective-api
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Perspective API site
-- product: pyrit
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: GitHub
-- product: qwen3guard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: shieldgemma
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
-- product: wildguard
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card and the project repository
-- product: zentropi-cope
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the HF model card
