--- sources/provenance/telemetry_observability.yaml
@@ -1,153 +1,0 @@
-# Prose closeout for telemetry_observability. 26 of 26 ready.
-#
-# All 26 were canonical before this pass; 21 still failed prose compliance. Sixteen were caught
-# by the ban detector and FIVE WERE NOT, which is why the complete prose of all 26 was read
-# rather than only the flagged ones:
-#
-#   agentops    comments named two licenses by name, which the detector's pattern missed
-#   azure-ai-foundry-observability  116-word description ending "it is the default
-#               observability layer for Azure-hosted GenAI workloads"
-#   vertex-ai-model-observability   107-word description, "Differentiated as the default
-#               zero-config observability surface"
-#   helicone    "Positioned as the lowest-friction entry point for LLM monitoring"
-#   weave       "Leverages W&B's established ML experiment tracking platform and enterprise
-#               customer base" - an adoption claim in a description
-#
-# The detector produces candidates. It does not define the scope.
-#
-# ## The dominant violation here was the openness verdict in the first clause
-#
-# Nine records opened with "Open-source X" or "Proprietary X" as the product's category noun -
-# langfuse, laminar, langwatch, mlflow, opik, phoenix, trulens, langtrace, langsmith. For this
-# category that reads naturally, because these vendors market themselves that way. It is still
-# the openness axis's content, and it is still a second copy with no evidence behind it.
-#
-# What replaced it is the deployment fact, which is architecture: "deploys from the repository's
-# docker-compose file", "runs on infrastructure you operate", "offered as managed cloud,
-# bring-your-own-cloud or a self-hosted enterprise deployment".
-#
-# ## agenta: the description described a product that no longer exists
-#
-# Agenta 2.0 repositioned from an LLMOps prompt and evaluation platform to a no-code agent
-# workspace. The capability axis was re-scored 4 -> 3 on 2026-08-14 with its peer anchor REMOVED
-# rather than repointed, because nothing else in this category is an agent workspace to compare
-# it against. The description still described the old product; it now describes the new one.
-#
-# The score note's open question is recorded and not resolved here: a no-code agent workspace
-# sits closer to orchestration_agents than to telemetry, and that is a recategorization for a
-# human rather than a re-score.
-#
-# ## Evidence-gap footnotes kept, because they are about our reading
-#
-# `helicone` records that the pricing page lists enterprise SSO the repository does not
-# implement. `laminar` records that its PyPI count is mirror-inflated and discounted rather than
-# read straight. `langsmith` records that its declared package is a hard dependency of
-# langchain-core, so the download count measures LangChain installs. `openllmetry` records that
-# the band was read on a different package than the one declared. `mlflow` records that it
-# declares no repository, so adoption routes through PyPI alone. `whylabs` records that the
-# company has discontinued operations.
-#
-- product: agenta
-  resolved_state: canonical
-  verified: '2026-08-14'
-  document: the Agenta documentation and repository
-- product: agentops
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the AgentOps-AI/agentops repository and its README
-- product: arize
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: arize
-- product: azure-ai-foundry-observability
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Microsoft Learn observability concept page
-- product: braintrust
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: braintrust
-- product: confident-ai
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: confident-ai
-- product: datadog-llm-observability
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Datadog LLM Observability documentation
-- product: dynatrace-ai-observability
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Dynatrace AI observability documentation
-- product: galileo
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: galileo
-- product: helicone
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Helicone/helicone README
-- product: laminar
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the lmnr-ai/lmnr repository and the lmnr PyPI project page
-- product: langfuse
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the langfuse/langfuse repository and the Langfuse observability docs
-- product: langsmith
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: langchain
-- product: langtrace
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Scale3-Labs/langtrace repository
-- product: langwatch
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the langwatch/langwatch repository
-- product: mlflow
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the mlflow/mlflow LICENSE file, the repository and mlflow
-- product: new-relic-ai-monitoring
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the New Relic AI Monitoring product page
-- product: openlit
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the openlit/openlit repository
-- product: openllmetry
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the traceloop/openllmetry repository
-- product: opik
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the comet-ml/opik repository and the opik PyPI project page
-- product: phoenix
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Arize-ai/phoenix repository and the Arize Phoenix product page
-- product: promptlayer
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the MagnivOrg/prompt-layer-library repository
-- product: trulens
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the truera/trulens repository
-- product: vertex-ai-model-observability
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the Agent Engine tracing documentation
-- product: weave
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the wandb/weave README
-- product: whylabs
-  resolved_state: canonical
-  verified: '2026-08-13'
-  document: the WhyLabs LangKit page and the whylabs/whylogs repository
