+++ tests/test_product_prose.py
@@ -30,7 +30,10 @@ person about — read the match in context before fixing it.
 import re
 from pathlib import Path
 
+import pytest
 import yaml
+
+from build.product_prose import census, classify
 
 ROOT = Path(__file__).resolve().parents[1]
 
@@ -93,3 +96,80 @@ def test_the_scan_actually_walks_the_corpus():
     """
     slugs = [slug for slug, _ in _descriptions()]
     assert len(slugs) > 400, f"only walked {len(slugs)} products; the glob has drifted"
+
+
+# --- the verification line -------------------------------------------------------------
+#
+# `docs/reference/product-copy.md` standardizes it on `Verified <date> via <document>.` The
+# 2026-08 closeout brought all 472 products to that form (tag `baseline-472-2026-08-16`).
+# What follows keeps them there: a new product, or an edit to an existing one, can reintroduce
+# any of the four unresolved states, and `sweep_status` reports the state it is told.
+
+
+@pytest.mark.parametrize(
+    "comments, expected",
+    [
+        ("Verified 2026-08-13 via the project README.", "canonical"),
+        ("Verified live 2026-08-13 via the project README.", "named_noncanonical"),
+        ("Verified live 2026-08-13 on huggingface.co/datasets/x.", "named_noncanonical"),
+        ("Verified live 2026-08-13 against the DATASHEET.md.", "named_noncanonical"),
+        # The substantive class: names how somebody looked, not what settled it.
+        ("Verified live 2026-08-13 via primary sources.", "generic"),
+        ("Verified 2026-08-13 via the primary sources.", "generic"),
+        ("Verified 2026-08-13 via substitute sources - hailo.ai 403s.", "generic"),
+        # Dated, not a method, and naming nothing. Defaulting these to `named_noncanonical`
+        # promised a mechanical fix that does not exist.
+        ("Verified 2026-08-13.", "ambiguous_noncanonical"),
+        ("Verified 2026-08-13 via .", "ambiguous_noncanonical"),
+        ("Verified live 2026-08-13, but only the adoption axis could be re-derived.",
+         "ambiguous_noncanonical"),
+        ("verified live 2026-08-13 via primary sources; v1.2.", "generic"),
+        ("Verified live 2026-08-13 via primary-source research.", "generic"),
+        ("Some prose with no verification at all.", "missing"),
+    ],
+)
+def test_classification(comments, expected):
+    assert classify(comments)[0] == expected
+
+
+@pytest.mark.parametrize(
+    "comments",
+    [
+        # Canonical in SHAPE, naming a method. `generic` has to outrank `canonical` or the
+        # check grades on shape while the thing it exists to find sits in the payload.
+        "Verified 2026-08-13 via primary sources.",
+        "Verified 2026-08-13 via research.",
+        "Verified 2026-08-13 via web search.",
+        # The live case: removing `live` from this would have promoted it into canonical
+        # form. Four hardware records were rewritten that way before this was caught.
+        "Verified 2026-08-13 via substitute sources - hailo.ai answers 403 to every request.",
+    ],
+)
+def test_canonical_form_naming_a_method_is_still_generic(comments):
+    assert classify(comments)[0] == "generic"
+
+
+def test_every_product_names_a_document_a_reader_can_reopen():
+    """The corpus-wide invariant, and the one this file exists to hold.
+
+    A failure names the products and their state. `generic` needs a document supplied;
+    `named_noncanonical` needs the wording brought to `via`; `ambiguous_noncanonical` and
+    `missing` need somebody to read the record.
+    """
+    by_state = census()
+    unresolved = {
+        state: slugs for state, slugs in by_state.items() if state != "canonical"
+    }
+    assert not unresolved, (
+        "every verification line must read `Verified <date> via <document>.` — see "
+        f"docs/reference/product-copy.md:\n{unresolved}"
+    )
+
+
+def test_the_census_walks_the_whole_corpus():
+    """Same guard as above, on the other instrument: an empty census satisfies the invariant
+    trivially, and two corpus walks in this repo's history silently narrowed."""
+    by_state = census()
+    assert sum(len(v) for v in by_state.values()) == len(list(_descriptions()))
+    assert set(by_state) <= {"canonical", "named_noncanonical", "ambiguous_noncanonical",
+                             "generic", "missing"}
